Move the `developer` page to the app router
Move shared components fromt the `inspect` page to the
shared folder.
