+++ app/src/app/scat-cs/inspect/terms-of-use-check.tsx
@@ -6,12 +6,12 @@
 
 import { useState } from "react";
 
+import { HowToCite } from "@/shared/how-to-cite";
 import { Button, Center, Modal } from "@mantine/core";
 import { IconLicense } from "@tabler/icons-react";
 import { DOWNLOAD_COOKIE_NAME } from "../../../shared/download";
 import { TermsOfUse } from "../../../shared/terms-of-use";
 import { ButtonClipboard } from "./button-clipboard";
-import { HowToCite } from "./how-to-cite";
 import { FormattedReference } from "./types";
 
 interface Props {
