+++ app/src/shared/reference-list.module.css
@@ -1,0 +1,11 @@
+/*
+ * SPDX-FileCopyrightText: LXCat team
+ *
+ * SPDX-License-Identifier: AGPL-3.0-or-later
+ */
+
+.scrollableTable {
+    :global(.mantine-ScrollArea-viewport) {
+        max-height: 300px;
+    }
+}
