+++ app/src/shared/reference-list.tsx
@@ -2,10 +2,10 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
+import { Reference } from "@/app/scat-cs/inspect/reference";
+import { FormattedReference } from "@/app/scat-cs/inspect/types";
 import { DataTable } from "mantine-datatable";
-import classes from "./inspect.module.css";
-import { Reference } from "./reference";
-import { FormattedReference } from "./types";
+import classes from "./reference-list.module.css";
 
 export const ReferenceList = ({
   references,
