+++ app/src/shared/terms-of-use-developer.tsx
@@ -2,19 +2,23 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
+import { Fieldset, Space, Text, Title } from "@mantine/core";
 import { HowToCite } from "./how-to-cite";
 import { TermsOfUse } from "./terms-of-use";
 
 export const TermsOfUseCheckForDeveloper = () => {
   return (
-    <fieldset>
-      <legend>Terms of use</legend>
+    <Fieldset>
+      <Title order={2}>Terms of use</Title>
+      <Space h="sm" />
       <TermsOfUse />
       <p>
-        The data retrieved from API endpoints should be cited by their
-        references.
+        <Text>
+          The data retrieved from API endpoints should be cited using their
+          references.
+        </Text>
       </p>
       <HowToCite references={[]} />
-    </fieldset>
+    </Fieldset>
   );
 };
