+++ app/src/shared/terms-of-use.tsx
@@ -2,20 +2,20 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
+import { Text } from "@mantine/core";
+
 export const TermsOfUse = () => {
   return (
     <>
-      <p>
+      <Text>
         Users acknowledge understanding that LXCat is a community-based project
         with open-access databases being freely provided by individual
-        contributors.
-      </p>
-      <p>
-        <b>
+        contributors.{" "}
+        <Text fw={700} span>
           Proper referencing of material retrieved from this site is essential
-          for the survival of the project.
-        </b>
-      </p>
+          for the longevity of the project.
+        </Text>
+      </Text>
     </>
   );
 };
